% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bku.R
\name{bku_fb}
\alias{bku_fb}
\title{Compute the sensitivity bound with formal benchmarking}
\usage{
bku_fb(y, m, a, c, j, ky_bound, km_bound, ka_bound)
}
\arguments{
\item{y}{a vector of the outcome variable}

\item{m}{a matrix of the mediator variables}

\item{a}{a vector of the exposure variable}

\item{c}{a matrix of the covariates}

\item{j}{a number of index to specify the refernece covaraite cj}

\item{ky_bound}{a number of the upper bound for the relative strength of the outcome-confounder correlation}

\item{km_bound}{a number of the upper bound for the relative strength of the mediator-confounder correlation}

\item{ka_bound}{a number of the upper bound for the relative strength of the exposure-confounder correlation}
}
\value{
the worst point estimate for the direct effect, the worst point estimate for the indirect effect, the worst t statistic for the direct effect, and the worst t statistic for the indirect effect
}
\description{
Compute the sensitivity bound with formal benchmarking
}
\examples{
set.seed(1234)
library(MASS)
n = 200
a = rnorm(n)
c = cbind(rnorm(n), rnorm(n))
m = mvrnorm(n, mu = c(0, 0), Sigma = matrix(c(1,0.2,0.2,0.8), 2, 2))
m = m + cbind(2.5 * a, 1.5 * a)
y = a + 0.2 * m[, 1] + 0.25 * m[, 2] + rnorm(n)
result = bku_fb(y, m, a, c, 1, 0.2, 0.2, 0.2)
}
