% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bku.R
\name{bku_indirect}
\alias{bku_indirect}
\title{Compute the sensitivity bound for the indirect effect with prespecified sensitivity parameters}
\usage{
bku_indirect(y, m, a, c = NULL, Ry, Rm, Ra)
}
\arguments{
\item{y}{a vector of the outcome variable}

\item{m}{a matrix of the mediator variables}

\item{a}{a vector of the exposure variable}

\item{c}{a matrix of the covariates}

\item{Ry}{a number of the R parameter for the outcome-confounder correlation}

\item{Rm}{a vector of the R parameter for the mediator-confounder correlation}

\item{Ra}{a number of the R parameter for the exposure-confounder correlation}
}
\value{
the point estimate, the standard error, and the t statistic
}
\description{
Compute the sensitivity bound for the indirect effect with prespecified sensitivity parameters
}
\examples{
set.seed(1234)
library(MASS)
n = 200
a = rnorm(n)
c = as.matrix(rnorm(n))
m = mvrnorm(n, mu = c(0, 0), Sigma = matrix(c(1,0.2,0.2,0.8), 2, 2))
m = m + cbind(2.5 * a, 1.5 * a)
y = a + 0.2 * m[, 1] + 0.25 * m[, 2] + rnorm(n)
result = bku_indirect(y, m, a, c, 0.1, c(0.2, 0.1), 0.05)
}
