% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bku.R
\name{bku_rv}
\alias{bku_rv}
\title{Compute the robustness values for both direct and indirect effects}
\usage{
bku_rv(
  y,
  m,
  a,
  c = NULL,
  randomized = F,
  rho_values = c(1:99)/100,
  vector_u = F,
  algorithm = "NLOPT_GN_ORIG_DIRECT_L"
)
}
\arguments{
\item{y}{a vector of the outcome variable}

\item{m}{a matrix of the mediator variables}

\item{a}{a vector of the exposure variable}

\item{c}{a matrix of the covariates}

\item{randomized}{if the exposure is randomized}

\item{rho_values}{the values of rho's to be considered for the grid search}

\item{vector_u}{if the unmeasured confounder is a vector}

\item{algorithm}{the optimization algorithm in the R package "nloptr"}
}
\value{
a summary table, the worst point estimates and the worst t statistics given each value of rho.
}
\description{
Compute the robustness values for both direct and indirect effects
}
\examples{
set.seed(1234)
library(MASS)
n = 200
a = rnorm(n)
c = as.matrix(rnorm(n))
m = mvrnorm(n, mu = c(0, 0), Sigma = matrix(c(1,0.2,0.2,0.8), 2, 2))
m = m + cbind(2.5 * a, 1.5 * a)
y = a + 0.2 * m[, 1] + 0.25 * m[, 2] + rnorm(n)
result = bku_rv_indirect(y, m, a, c)
}
