% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bku.R
\name{bku_worst_direct}
\alias{bku_worst_direct}
\title{Compute the worst point estimate and t statistic for the direct effect under upper bound of the R parameters}
\usage{
bku_worst_direct(y, m, a, c = NULL, rho_y, rho_m, rho_a)
}
\arguments{
\item{y}{a vector of the outcome variable}

\item{m}{a matrix of the mediator variables}

\item{a}{a vector of the exposure variable}

\item{c}{a matrix of the covariates}

\item{rho_y}{a number of upper bound of the R parameter for the outcome-confounder correlation}

\item{rho_m}{a number of upper bound of the R parameter for the mediator-confounder correlation}

\item{rho_a}{a number of upper bound of the R parameter for the exposure-confounder correlation}
}
\value{
the worst point estimate, and the worst t statistic
}
\description{
Compute the worst point estimate and t statistic for the direct effect under upper bound of the R parameters
}
\examples{
set.seed(1234)
library(MASS)
n = 200
a = rnorm(n)
c = as.matrix(rnorm(n))
m = mvrnorm(n, mu = c(0, 0), Sigma = matrix(c(1,0.2,0.2,0.8), 2, 2))
m = m + cbind(2.5 * a, 1.5 * a)
y = a + 0.2 * m[, 1] + 0.25 * m[, 2] + rnorm(n)
result = bku_worst_direct(y, m, a, c, 0.1, 0.1, 0.1)
}
