% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{bkuvm}
\alias{bkuvm}
\title{Sensitivity analysis for the Baron-Kenny approach with multiple mediators}
\usage{
bkuvm(
  y,
  m,
  a,
  c = NULL,
  r2.max = 0.5,
  randomized.a = FALSE,
  s1 = c(1, -1),
  s5 = c(1, -1),
  s6 = c(1, -1),
  s7 = c(1, -1),
  B = 500,
  N = 10,
  shortcut = FALSE,
  progress.display = TRUE
)
}
\arguments{
\item{y}{a vector of the outcome variable}

\item{m}{a matrix of the mediator variables}

\item{a}{a vector of the treatment variable}

\item{c}{a vector or matrix of the covariates}

\item{r2.max}{the maximum r^2 to be considered for the grid search}

\item{randomized.a}{if the treatment is randomized}

\item{s1}{(optional) the pre-specified sign S1; see Zhang and Ding (2022)}

\item{s5}{(optional) the pre-specified sign S5; see Zhang and Ding (2022)}

\item{s6}{(optional) the pre-specified sign S6; see Zhang and Ding (2022)}

\item{s7}{(optional) the pre-specified sign S7; see Zhang and Ding (2022)}

\item{B}{the number of bootstrap resampling}

\item{N}{the number of Monte Carlo sampling to generate \eqn{R_{m\sim u\mid a,c}^2} given fixed \eqn{R_{u\sim m\mid a,c}^2}}

\item{shortcut}{if use the shortcut approach}

\item{progress.display}{if display the progress}
}
\value{
a summary table, and the point estimates, standard errors, t statistics
of the direct and indirect effects with sensitivity parameters generated by
the grid search.
}
\description{
Sensitivity analysis for the Baron-Kenny approach with multiple mediators
}
\examples{
set.seed(1234)
a = rnorm(1000)
m1 = 0.2 * a + rnorm(1000)
m2 = 0.3 * a + rnorm(1000)
m = cbind(m1, m2)
u = (sqrt(1/3.2)-sqrt(1/8)*0.5) * a + sqrt(1/8) * m1 + sqrt(1/8) * m2 + rnorm(1000)
y = 0.5 * u + rnorm(1000)
cor(a,u)^2
cor(lm(y~a+m)$residuals, lm(u~a+m)$residuals)^2
1 - var(lm(u~a+m)$residuals) / var(lm(u~a)$residuals)
result = bkuvm(y, m, a, shortcut = TRUE)
result$coefficients
}
